embedding-svc/requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
alembic==1.13.1
sqlalchemy==2.0.29

# ── Embeddings ───────────────────────────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0
open-clip-torch==2.24.0
Pillow==10.1.0
numpy==1.26.0
requests==2.31.0

# ── Vector store ─────────────────────────────────────────────────────────────
pinecone==5.1.0

# ── Config ───────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
